+++ ChangeLog.rst
@@ -6,6 +6,13 @@ Note worthy changes
 
 - IdP: Added support for the device authorization grant.
 
+Fixes
+-----
+
+- The recently added support for phone (SMS) authentication did fully integrate
+  with third-party provider signups. For example, whether or not the phone
+  number is required was not respected during signup. Fixed.
+
 
 65.9.0 (2025-06-01)
 *******************